:sparkles: :: 별점 Dto 추가

+++ src/main/java/com/hackathon/daechelinguide/domain/review/presentation/dto/response/ReviewResponseDto.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @Builder
 public class ReviewResponseDto {
-    private Integer star;
+    private double star;
     private String menu;
     private String date;
 }
